move string macros
